SDLTimer and implemented delta time!

+++ include/modules/D_Config.h
@@ -5,6 +5,7 @@
 #ifndef D_CONFIG_H
 #define D_CONFIG_H
 
+#include <stdint.h>
 #include <string>
 
 namespace Diamond {
@@ -18,15 +19,15 @@ namespace Diamond {
 		std::string game_name = "A Game Without a Name";
 
 		// Graphics settings
-		int window_width = 1280;
-		int window_height = 720;
+		int16_t window_width = 1280;
+		int16_t window_height = 720;
 		bool vsync = false;
 		RGBA bg_color = RGBA{0, 0, 0, 100};
 
 		// Audio settings
 		audio_channels_t audio_channels = D_STEREO;
-		int audio_out_freq = 44100; // in hertz
-		int audio_out_sample_size = 2048; // in bytes
+		int32_t audio_out_freq = 44100; // in hertz
+		int16_t audio_out_sample_size = 2048; // in bytes
 	};
 }
